chore: refactor and reorganize type definitions for Atlas service
